close resp body gracefully and replace drained res body with a copy
- replacing the drained res body allows the client to still consume
the response body in the regular fashion. Some valid use case for this
might be responses where server might reply with an error instead of
the expected struct to marshal..in this case the user might want to know
if the body of the reponse contained useful information that might help
with debugging.